+++ Playground/RxSwiftExtPlayground.playground/Pages/mapMany.xcplaygroundpage/Contents.swift
@@ -1,6 +1,6 @@
 /*:
  > # IMPORTANT: To use `RxSwiftExtPlayground.playground`, please:
- 
+
  1. Make sure you have [Carthage](https://github.com/Carthage/Carthage) installed
  1. Fetch Carthage dependencies from shell: `carthage bootstrap --platform ios`
  1. Build scheme `RxSwiftExtPlayground` scheme for a simulator target
@@ -11,37 +11,35 @@
 import RxSwift
 import RxSwiftExt
 
-/*:
- ## toSortedArray
- 
- The `toSortedArray` operator transforms an observable of comparables into an observable of ordered arrays.
- */
+example("mapMany") {
+    let numbers = Observable.of(1...10)
+    let strings = Observable.of(["RxSwift", "is" ,"awesome", "along", "with", "RxSwiftCommunity"])
 
-example("Ensure that only a sorted array is emitted") {
-    let sequenceStream = Observable.of(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12)
-    let array = Observable.of(10, 9, 12, 8, 4, 1, 1, 8, 14)
-
-    // Ascending order
-    array.toSortedArray()
-        .subscribe(onNext: { result in
-            print(result)
-        })
-    
-    array.toSortedArray(ascending: true)
-        .subscribe(onNext: { result in
-            print(result)
-        })
-    
-    // Descending order
-    sequenceStream.toSortedArray(by: >)
+    // Map many using a model initializer
+    numbers.mapMany(SomeModel.init)
         .subscribe(onNext: { result in
             print(result)
         })
 
-    array.toSortedArray(ascending: false)
+    // Map many with a transformation closure
+    numbers.mapMany { $0 * $0 }
         .subscribe(onNext: { result in
             print(result)
         })
+
+    strings.mapMany { $0.lowercased() }
+        .subscribe(onNext: { result in
+            print(result)
+        })
+
+    struct SomeModel: CustomStringConvertible {
+        let number: Int
+        var description: String { return "#\(number)" }
+
+        init(_ number: Int) {
+            self.number = number
+        }
+    }
 }
 
 //: [Next](@next)